Move `run` command functions to `pkg/run`.
